requirements.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.7.3
cbor2==5.4.6
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cytoolz==0.12.1
dacite==1.8.0
Deprecated==1.2.13
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fastapi==0.94.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.4
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.3
pycryptodome==3.17
pydantic==1.10.6
pyee==9.0.4
pymerkle==2.0.2
pyrsistent==0.19.3
pytz==2022.7.1
requests==2.28.2
rlp==2.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.26.0.post1
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==3.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
varint==1.0.2
web3==5.27.0
websockets==9.1
wrapt==1.15.0
yarl==1.8.2
